update user registration and update

+++ resources/views/admin/dashboard/allservices.blade.php
@@ -37,8 +37,10 @@
 										
 											 <div class="d-flex order-actions">
 											<a href="{{route('admin.edit-service',['id'=>$service_id])}}" class="bg-primary" style="color:white"><i class='bx bxs-edit'></i></a>
-
+											@if ($admin_data->user_type=="admin")
 												<a href="javascript:;"  onclick="DeleteService({{$value->service_id}})"  class="ms-3 bg-danger" style="color: white"><i class='bx bxs-trash'></i></a>
+												
+											@endif
 											 </div>
 										</td>
 									</tr>
